composition_repository/django_deovi/requirements.txt
# Temporarly pin bigtree with pandas extra requirement until
# django-deovi fixed bigtree requirement to enable the extra one
bigtree[pandas]==0.10.3
django-deovi[breadcrumbs]==0.6.1